+++ src/App/Fossa/Config/Analyze.hs
@@ -202,9 +202,11 @@ data AnalyzeCliOpts = AnalyzeCliOpts
   , analyzeSkipVSIGraphResolution :: [VSI.Locator]
   , analyzeBaseDir :: FilePath
   , analyzeDynamicGoAnalysisType :: GoDynamicTactic
+  , analyzePathDependencies :: Bool
   , analyzeForceFirstPartyScans :: Flag ForceFirstPartyScans
   , analyzeForceNoFirstPartyScans :: Flag ForceNoFirstPartyScans
   , analyzeIgnoreOrgWideCustomLicenseScanConfigs :: Flag IgnoreOrgWideCustomLicenseScanConfigs
+  , analyzeCustomFossaDepsFile :: Maybe FilePath
   }
   deriving (Eq, Ord, Show)
 
@@ -233,6 +235,7 @@ data AnalyzeConfig = AnalyzeConfig
   , overrideDynamicAnalysis :: OverrideDynamicAnalysisBinary
   , firstPartyScansFlag :: FirstPartyScansFlag
   , grepOptions :: GrepOptions
+  , customFossaDepsFile :: Maybe FilePath
   }
   deriving (Eq, Ord, Show, Generic)
 
@@ -242,6 +245,7 @@ instance ToJSON AnalyzeConfig where
 data ExperimentalAnalyzeConfig = ExperimentalAnalyzeConfig
   { allowedGradleConfigs :: Maybe (Set Text)
   , useV3GoResolver :: GoDynamicTactic
+  , resolvePathDependencies :: Bool
   }
   deriving (Eq, Ord, Show, Generic)
 
@@ -280,9 +284,11 @@ cliParser =
     <*> many skipVSIGraphResolutionOpt
     <*> baseDirArg
     <*> experimentalUseV3GoResolver
+    <*> experimentalAnalyzePathDependencies
     <*> flagOpt ForceFirstPartyScans (long "experimental-force-first-party-scans" <> help "Force first party scans")
     <*> flagOpt ForceNoFirstPartyScans (long "experimental-block-first-party-scans" <> help "Block first party scans. This can be used to forcibly turn off first-party scans if your organization defaults to first-party scans.")
     <*> flagOpt IgnoreOrgWideCustomLicenseScanConfigs (long "ignore-org-wide-custom-license-scan-configs" <> help "Ignore custom-license scan configurations for your organization. These configurations are defined in the \"Integrations\" section of the Admin settings in the FOSSA web app")
+    <*> optional (strOption (long "fossa-deps-file" <> help "Path to fossa-deps file including filename (default: fossa-deps.{yaml|yml|json})"))
 
 data GoDynamicTactic
   = GoModulesBasedTactic
@@ -304,6 +310,14 @@ experimentalUseV3GoResolver =
       <> help
         ( coloredText Red "DEPRECATED: This is now default and will be removed in the future."
             <> " For Go: generate a graph of module deps based on package deps. This will be the default in the future."
+        )
+
+experimentalAnalyzePathDependencies :: Parser Bool
+experimentalAnalyzePathDependencies =
+  switch $
+    long "experimental-analyze-path-dependencies"
+      <> help
+        ( "License scan dependencies sourced from file system, as indicated in manifest files. This will be enabled by default in the future."
         )
 
 vendoredDependencyModeOpt :: Parser ArchiveUploadType
@@ -406,6 +420,8 @@ mergeStandardOpts maybeConfig envvars cliOpts@AnalyzeCliOpts{..} = do
       vendoredDepsOptions = collectVendoredDeps maybeConfig cliOpts
       dynamicAnalysisOverrides = OverrideDynamicAnalysisBinary $ envCmdOverrides envvars
       grepOptions = collectGrepOptions maybeConfig cliOpts
+      customFossaDepsFile = analyzeCustomFossaDepsFile
+
   firstPartyScansFlag <-
     case (fromFlag ForceFirstPartyScans analyzeForceFirstPartyScans, fromFlag ForceNoFirstPartyScans analyzeForceNoFirstPartyScans) of
       (True, True) -> fatalText "You provided both the --experimental-force-first-party-scans and --experimental-block-first-party-scans flags. Only one of these flags may be used"
@@ -429,6 +445,7 @@ mergeStandardOpts maybeConfig envvars cliOpts@AnalyzeCliOpts{..} = do
     <*> pure dynamicAnalysisOverrides
     <*> pure firstPartyScansFlag
     <*> pure grepOptions
+    <*> pure customFossaDepsFile
 
 collectFilters ::
   ( Has Diagnostics sig m
@@ -455,13 +472,14 @@ collectCLIFilters AnalyzeCliOpts{..} =
     (comboExclude analyzeExcludeTargets analyzeExcludePaths)
 
 collectExperimental :: Maybe ConfigFile -> AnalyzeCliOpts -> ExperimentalAnalyzeConfig
-collectExperimental maybeCfg AnalyzeCliOpts{analyzeDynamicGoAnalysisType = goDynamicAnalysisType} =
+collectExperimental maybeCfg AnalyzeCliOpts{analyzeDynamicGoAnalysisType = goDynamicAnalysisType, analyzePathDependencies = shouldAnalyzePathDependencies} =
   ExperimentalAnalyzeConfig
     ( fmap
         gradleConfigsOnly
         (maybeCfg >>= configExperimental >>= gradle)
     )
     goDynamicAnalysisType
+    shouldAnalyzePathDependencies
 
 collectVendoredDeps ::
   ( Has Diagnostics sig m