+++ src/App/Fossa/Container/Sources/DockerArchive.hs
@@ -179,6 +179,7 @@ analyzeLayer systemDepsOnly filters capabilities osInfo layerFs tarball = do
       ExperimentalAnalyzeConfig
         Nothing
         GoModulesBasedTactic -- Discovery is the same for both module and package centric analysis.
+        False -- Discovery has no consequence from path dependency analysis config
     toSourceUnit :: [DiscoveredProjectScan] -> [SourceUnit]
     toSourceUnit =
       map (Srclib.projectToSourceUnit False)
@@ -321,6 +322,7 @@ listTargetLayer capabilities osInfo layerFs tarball layerType = do
       ( ExperimentalAnalyzeConfig
           Nothing
           GoModulesBasedTactic -- Targets aren't different between package/module centric analysis for Go.
+          False -- Targets are not impacted by path dependencies.
       )
     . runReader (mempty :: AllFilters)
     $ run