+++ src/App/Fossa/ManualDeps.hs
@@ -23,6 +23,7 @@ import App.Fossa.ArchiveUploader (archiveUploadSourceUnit)
 import App.Fossa.Config.Analyze (
   VendoredDependencyOptions (..),
  )
+import App.Fossa.Config.Common (validateFile)
 import App.Fossa.LicenseScanner (licenseScanSourceUnit)
 import App.Fossa.VendoredDependency (
   VendoredDependency (..),
@@ -64,6 +65,7 @@ import Path (Abs, Dir, File, Path, mkRelFile, (</>))
 import Path.Extra (tryMakeRelative)
 import Srclib.Converter (depTypeToFetcher)
 import Srclib.Types (AdditionalDepData (..), Locator (..), SourceRemoteDep (..), SourceUnit (..), SourceUnitBuild (..), SourceUnitDependency (SourceUnitDependency), SourceUserDefDep (..), someBaseToOriginPath)
+import System.FilePath (takeExtension)
 import Types (ArchiveUploadType (..), GraphBreadth (..))
 
 data FoundDepsFile
@@ -80,16 +82,37 @@ analyzeFossaDepsFile ::
   , Has Exec sig m
   ) =>
   Path Abs Dir ->
+  Maybe FilePath ->
   Maybe ApiOpts ->
   VendoredDependencyOptions ->
   m (Maybe SourceUnit)
-analyzeFossaDepsFile root maybeApiOpts vendoredDepsOptions = do
-  maybeDepsFile <- findFossaDepsFile root
+analyzeFossaDepsFile root maybeCustomFossaDepsPath maybeApiOpts vendoredDepsOptions = do
+  maybeDepsFile <-
+    case maybeCustomFossaDepsPath of
+      Nothing -> findFossaDepsFile root
+      Just filePath -> retrieveCustomFossaDepsFile filePath
   case maybeDepsFile of
     Nothing -> pure Nothing
     Just depsFile -> do
       manualDeps <- context "Reading fossa-deps file" $ readFoundDeps depsFile
       context "Converting fossa-deps to partial API payload" $ Just <$> toSourceUnit root depsFile manualDeps maybeApiOpts vendoredDepsOptions
+
+retrieveCustomFossaDepsFile ::
+  ( Has Diagnostics sig m
+  , Has (Lift IO) sig m
+  , Has ReadFS sig m
+  ) =>
+  FilePath ->
+  m (Maybe FoundDepsFile)
+retrieveCustomFossaDepsFile fossaDepsPath = do
+  let extension = takeExtension fossaDepsPath
+  file <- validateFile fossaDepsPath
+
+  case extension of
+    ".yml" -> pure $ Just $ ManualYaml file
+    ".yaml" -> pure $ Just $ ManualYaml file
+    ".json" -> pure $ Just $ ManualJSON file
+    _ -> fatalText $ "Expected <name-of-file>.{yml|yaml|json} but received: " <> toText fossaDepsPath
 
 findAndReadFossaDepsFile ::
   ( Has Diagnostics sig m